EQ Appearance Changes
Should be easy to identify the ship number now. Let me know if I screwed up the layout.
Its suppose to look like this: https://i.imgur.com/F2WqhJl.png

Moved the help info out of the footer space, so it can be used for having the timestamp there.